printer-agent-server/requirements.txt
fastapi===0.115.12
uvicorn[standard]===0.38.0
pydantic===2.11.5
nuitka===2.8.6
python-escpos===3.1
pyusb===1.3.1
jinja2===3.0.3
python-multipart===0.0.20
zeroconf===0.148.0
